scripts/logo-scrapper/requirements.txt
requests
beautifulsoup4
pillow
cairosvg    # optional but strongly recommended for SVG support
